RequirementMachine: Split off RuleBuilder.{cpp,h} from RequirementLowering.{cpp,h}
