+++ lib/AST/CMakeLists.txt
@@ -95,6 +95,7 @@ add_swift_host_library(swiftAST STATIC
   RequirementMachine/RewriteLoop.cpp
   RequirementMachine/RewriteSystem.cpp
   RequirementMachine/Rule.cpp
+  RequirementMachine/RuleBuilder.cpp
   RequirementMachine/SimplifySubstitutions.cpp
   RequirementMachine/Symbol.cpp
   RequirementMachine/Term.cpp
