+++ lib/AST/RequirementMachine/RequirementLowering.h
@@ -1,4 +1,4 @@
-//===--- RequirementLowering.h - Building rules from requirements ---------===//
+//===--- RequirementLowering.h - Requirement inference and desugaring -----===//
 //
 // This source file is part of the Swift.org open source project
 //
@@ -19,10 +19,6 @@
 #include "llvm/ADT/SmallVector.h"
 #include <vector>
 #include "Diagnostics.h"
-#include "RewriteContext.h"
-#include "Rule.h"
-#include "Symbol.h"
-#include "Term.h"
 
 namespace llvm {
   class raw_ostream;
@@ -38,7 +34,7 @@ class Requirement;
 namespace rewriting {
 
 // Entry points used by AbstractGenericSignatureRequest and
-// InferredGenericSignatureRequest; see RequiremetnLowering.cpp for
+// InferredGenericSignatureRequest; see RequirementLowering.cpp for
 // documentation
 // comments.
 
@@ -62,88 +58,6 @@ void realizeInheritedRequirements(TypeDecl *decl, Type type,
 bool diagnoseRequirementErrors(ASTContext &ctx,
                                ArrayRef<RequirementError> errors,
                                bool allowConcreteGenericParams);
-
-/// A utility class for bulding rewrite rules from the top-level requirements
-/// of a generic signature.
-///
-/// This also collects requirements from the transitive closure of all protocols
-/// appearing on the right hand side of conformance requirements.
-struct RuleBuilder {
-  RewriteContext &Context;
-
-  /// The transitive closure of all protocols appearing on the right hand
-  /// side of conformance requirements.
-  llvm::DenseSet<const ProtocolDecl *> &ReferencedProtocols;
-
-  /// A subset of the above in insertion order, consisting of the protocols
-  /// whose rules we are going to import.
-  ///
-  /// If this is a rewrite system built from a generic signature, this vector
-  /// contains all elements in the above set.
-  ///
-  /// If this is a rewrite system built from a strongly connected component
-  /// of the protocol, this vector contains all elements in the above set
-  /// except for the protocols belonging to the component representing the
-  /// rewrite system itself; those protocols are added directly instead of
-  /// being imported.
-  std::vector<const ProtocolDecl *> ProtocolsToImport;
-
-  /// The rules representing a complete rewrite system for the above vector,
-  /// pulled in by collectRulesFromReferencedProtocols().
-  std::vector<Rule> ImportedRules;
-
-  /// New rules to add which will be marked 'permanent'. These are rules for
-  /// introducing associated types, and relationships between layout,
-  /// superclass and concrete type symbols. They are not eliminated by
-  /// homotopy reduction, since they are always added when the rewrite system
-  /// is built.
-  std::vector<std::pair<MutableTerm, MutableTerm>> PermanentRules;
-
-  /// New rules derived from requirements written by the user, which can be
-  /// eliminated by homotopy reduction.
-  std::vector<std::tuple<MutableTerm, MutableTerm, Optional<unsigned>>>
-      RequirementRules;
-
-  /// Requirements written in source code. The requirement ID in the above
-  /// \c RequirementRules vector is an index into this array.
-  std::vector<StructuralRequirement> WrittenRequirements;
-
-  /// Enables debugging output. Controlled by the -dump-requirement-machine
-  /// frontend flag.
-  unsigned Dump : 1;
-
-  /// Used to ensure the initWith*() methods are only called once.
-  unsigned Initialized : 1;
-
-  RuleBuilder(RewriteContext &ctx,
-              llvm::DenseSet<const ProtocolDecl *> &referencedProtocols)
-      : Context(ctx), ReferencedProtocols(referencedProtocols) {
-    Dump = ctx.getASTContext().LangOpts.DumpRequirementMachine;
-    Initialized = 0;
-  }
-
-  void initWithGenericSignatureRequirements(ArrayRef<Requirement> requirements);
-  void initWithWrittenRequirements(ArrayRef<StructuralRequirement> requirements);
-  void initWithProtocolSignatureRequirements(ArrayRef<const ProtocolDecl *> proto);
-  void initWithProtocolWrittenRequirements(ArrayRef<const ProtocolDecl *> proto);
-  void initWithConditionalRequirements(ArrayRef<Requirement> requirements,
-                                       ArrayRef<Term> substitutions);
-  void addReferencedProtocol(const ProtocolDecl *proto);
-  void collectRulesFromReferencedProtocols();
-
-private:
-  void addPermanentProtocolRules(const ProtocolDecl *proto);
-  void addAssociatedType(const AssociatedTypeDecl *type,
-                         const ProtocolDecl *proto);
-  void addRequirement(const Requirement &req,
-                      const ProtocolDecl *proto,
-                      Optional<ArrayRef<Term>> substitutions=None,
-                      Optional<unsigned> requirementID=None);
-  void addRequirement(const StructuralRequirement &req,
-                      const ProtocolDecl *proto);
-  void addTypeAlias(const ProtocolTypeAlias &alias,
-                    const ProtocolDecl *proto);
-};
 
 // Defined in ConcreteContraction.cpp.
 bool performConcreteContraction(
